closingsales/requirements.txt
arrow==0.14.2
dateparser==0.7.1
Django==2.1.3
django-debug-toolbar==1.11
django-extensions==2.1.9
django-widget-tweaks==1.4.5
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
moment==0.8.2
pbr==5.2.0
Pillow==6.0.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.7
regex==2019.6.8
six==1.12.0
sqlparse==0.3.0
stevedore==1.30.1
times==0.7
tzlocal==1.5.1
virtualenv==16.0.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.5
